refactor: Move ImageSetting enum into its own file
