+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/settings/ImageSetting.kt
@@ -1,0 +1,11 @@
+package io.github.pandier.intellijdiscordrp.settings
+
+enum class ImageSetting(
+    private val friendlyName: String
+) {
+    APPLICATION("Application"),
+    FILE("File");
+
+    override fun toString(): String =
+        friendlyName
+}
